requirements.txt
click==6.7
fuzzywuzzy==0.14.0
PyYAML==3.12
python-Levenshtein==0.12.0
unicodecsv==0.14.1
